add prompts for easier mobile use
