+++ polygon-draw/src/app/store.js
@@ -1,0 +1,5 @@
+import { configureStore } from "@reduxjs/toolkit";
+
+export default configureStore({
+  reducer: {},
+});
