+++ polygon-draw/src/styles/main.scss
@@ -1,0 +1,24 @@
+html,
+body {
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  background-color: black;
+  overflow: hidden;
+
+  //   height: 100vh;
+  //   width: 100vw;
+}
+
+#root {
+  height: 100%;
+}
+
+.container {
+  background-color: black;
+  height: 100%;
+  width: 100%;
+}
