+++ polygon-draw/src/styles/p5sketch.scss
@@ -1,0 +1,3 @@
+.react-p5 {
+  border: 5px dotted white;
+}
