+++ polygon-draw/src/styles/prompt.scss
@@ -1,0 +1,38 @@
+.prompt-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  color: white;
+  height: 100%;
+  width: 100%;
+
+  .prompt-header {
+    text-align: center;
+  }
+
+  .prompt-done-btn-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    vertical-align: middle;
+    justify-content: center;
+    flex: 1;
+    width: 100%;
+
+    .prompt-done-btn:active {
+      background-color: white;
+      color: black;
+      cursor: pointer;
+    }
+
+    .prompt-done-btn {
+      border: 3px solid white;
+      background-color: black;
+      color: white;
+      height: 40%;
+      width: 50%;
+      font-size: 2em;
+      font-weight: bold;
+    }
+  }
+}
